Regulated Afwijking data
There’s a bug related to the table-level expectation results. The results for table-level expectations differ slightly from the column-level results in the checkpoint_result, which causes the afwijking table not to be populated. As a result, the afwijking data cannot be retrieved because the table-level expectation results contain different some required fields in the checkpoint_result JSON. Arranged it.